requirements.txt
plotly
streamlit
pandas
matplotlib
scipy
fitter
numpy-financial
watchdog